requirements.txt
anyio==3.6.2
appdirs==1.4.4
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
et-xmlfile==1.1.0
fastapi==0.92.0
fonttools==4.38.0
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.0.0
jieba==0.42.1
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.0
numpy==1.24.2
openpyxl==3.1.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pydantic==1.10.5
pyee==8.2.2
pyparsing==3.0.9
pyppeteer==1.0.2
pyppeteer-stealth==2.7.4
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
wordcloud @ file:///C:/Users/29230/PycharmProjects/pythonProject/wordcloud-1.8.1-cp311-cp311-win_amd64.whl
zipp==3.13.0
